Implement a possible fix for flow
Due to the create-universal-package upgrade this package now exports an index.js.flow which flow will attempt to use for flowtyping. This file currently does not export any of the types within flow.js, or flow-typed/fusion-core.js, so this causes flow to fail. Ignoring src and the built .flow file should allow us to have flow working as it was before.

I'm not sure if this gets us anything for type safety when importing fusion-core, but it does not cause flow to fail for other plugins.